Implement basic inbuilt authorisation
This doesn't quite work yet, since the admin templates seem to clash
with the default templates used in the out-of-the-box views, and seem to
be taking precedence.

Tried a few things to remedy this, including reordering the installed
apps in the settings, but it's been a very long day and I'm too tired to
try anything else.

Will look at the LOGIN_REDIRECT_URL and its two pals later too.