Cull Django files not part of the application
July 2019 feels like a long time ago. Lots of pointless stuff was
included in the repo which, now wiser, I see is obviously not needed.
Now the repo reflects just the code which forms the Django application.

I'll save would-be ne'er-do-wells some time - the secret key in the
settings was never deployed anywhere except the development laptop.